requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.14.0
environs==9.5.0
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
grpcio==1.63.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.11
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.77
marshmallow==3.21.3
milvus-lite==2.4.7
milvus-model==0.2.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.18.0
openai==0.28.1
orjson==3.10.4
packaging==23.2
pandas==2.2.2
protobuf==5.27.1
psycopg2==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
pymilvus==2.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
sympy==1.12.1
tenacity==8.3.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
yarl==1.9.4
